Write an incomplete Bridge environment
I set up this repository in a way that should be compatible with
Open AI Gym (see README for link to writing an environment).
I didn't at all change the examples code (still the same as it was
in the tictactoe game).

I wrote all of the "bridge" game components and tried to fill out
the Environment components as best as I could - the code probably
does not run and is buggy but a lot of the infrastructure and logic
is now there.